+++ src/store/chat/slices/aiChat/actions/index.ts
@@ -2,16 +2,20 @@ import { StateCreator } from 'zustand/vanilla';
 
 import { ChatStore } from '@/store/chat/store';
 
-import { AIGenerateAction, generateAIChat } from './generateAIChat';
-import { AIGenerateV2Action, generateAIChatV2 } from './generateAIChatV2';
+import { ConversationControlAction, conversationControl } from './conversationControl';
+import { ConversationLifecycleAction, conversationLifecycle } from './conversationLifecycle';
 import { ChatMemoryAction, chatMemory } from './memory';
 import { ChatRAGAction, chatRag } from './rag';
+import { StreamingExecutorAction, streamingExecutor } from './streamingExecutor';
+import { StreamingStatesAction, streamingStates } from './streamingStates';
 
 export interface ChatAIChatAction
   extends ChatRAGAction,
     ChatMemoryAction,
-    AIGenerateAction,
-    AIGenerateV2Action {
+    ConversationLifecycleAction,
+    ConversationControlAction,
+    StreamingExecutorAction,
+    StreamingStatesAction {
   /**/
 }
 
@@ -22,7 +26,9 @@ export const chatAiChat: StateCreator<
   ChatAIChatAction
 > = (...params) => ({
   ...chatRag(...params),
-  ...generateAIChat(...params),
   ...chatMemory(...params),
-  ...generateAIChatV2(...params),
+  ...conversationLifecycle(...params),
+  ...conversationControl(...params),
+  ...streamingExecutor(...params),
+  ...streamingStates(...params),
 });
